Add Zpool.set_props and Zpool.set_props_version
The former has to look up the version from the pool to validate the
props.  The latter takes the version as a parameter.

Baking validation into the interface is somewhat awkward.  Validation is
really meant to check user input.  For code consumers, an interface
that assumes valid input may be useful...